feat: support committing recursive files

+++ pit/database.py
@@ -15,5 +15,7 @@ class Database:
 
     def store(self, obj: GitObject):
         path = self.objects_dir / obj.oid[:2] / obj.oid[2:]
+        if path.exists():
+            return
         path.parent.mkdir(parents=True, exist_ok=True)
         path.write_bytes(zlib.compress(obj.saved))
